Modelines, clean up vimdoc formatting and edit text
I almost messed up indentation when I started editing code before
realizing that it uses true tabs. So let's set some modelines so other
contributors won't mess it up too :-)

I just guessed at a 'tabstop' value error hoping the author is happy
with it :-)  It's less extreme than 8 or 2, so...

The doc file was (mostly) indented with spaces (but actually mixed), so
I converted the whole thing to spaces and set 'expandtab' in the
modeline.

Tried to clarify some language in the documentation.