Add option to use next window when excluded
I find it more useful to move to the next usable window rather than
show an error when the selected window contains an excluded filetype.

Although this is possible using custom functions, and I have done that
in my own vimrc, it seems like it might be a useful optional feature,
so this commit adds a g:qfenter_excluded_action option to allow that.

Set the option to 'next' to tell QFEnter to find the next usable window,
falling back to the quickfix default (split above) when no window found.

The fallback could be changed to show an error, but IMHO it is more
useful to open the file somewhere rather than simply error.